// src/app/pages/dashboard/quote/models/quotes-api-response.ts
export class QuotesApiResponse{
  status: number;
  data: [];
  message?: string;
  errors?: [];
}
export class srLine {
  approvedBy: string;
  approvedDate: string;
  bookingRefNo: string;
  channel: string;
  contactId: number;
  createdBy: number;
  createdDate: string;
  custmerId: number;
  customerOption: string;
  fareRules: string;
  fullfilledBy: string;
  fullfilledDate: string;
  id: number;
  internalPolicies: string;
  mergeId: number;
  productId: number;
  remarks: string;
  seqNo: number;
  srId: number;
  srLineId: number;
  statusCode: string;
  statusId: number;
  submittedBy: string;
  submittedDate: string;
}